+++ TirionFordringApp/src/progtech/entities/Unit.java
@@ -2,18 +2,18 @@ package progtech.entities;
 
 public abstract class Unit
 {
-    private byte hp;
-    private byte unitCost;
+    private short hp;
+    private short unitCost;
     private String displayName;
     
-    public Unit(byte hp, byte unitCost, String displayName)
+    public Unit(short hp, short unitCost, String displayName)
     {
         this.hp = hp;
         this.unitCost = unitCost;
         this.displayName = displayName;
     }
     
-    public byte getHp(){ return hp; }
-    public byte getUnitCost(){ return unitCost; }
+    public short getHp(){ return hp; }
+    public short getUnitCost(){ return unitCost; }
     public String getDisplayName(){ return displayName; }
 }
